Pinpoint - Switch UI over to using <app-route>
Previously, Pinpoint used the dashboard's pattern of having endpoints support both a GET (which rendered the html page) and a POST (which could be queried for json data). To add new pages, you'd need to go implement a new handler, and modify dispatcher.py to register a route.

Instead we can use Polymer's app-route element, which delegates routing to client side. In pinpoint.yaml we can specify that non-data endpoints are served the index.html file with an <app-location> element and <app-route> elements defined. This will control which pages are shown if a route matches, and can also delegate handling the rest of that sub route if we want to nest.

https://www.polymer-project.org/1.0/blog/routing
https://www.polymer-project.org/1.0/toolbox/routing

Demo:
https://dev-simonhatch-c7d2582c-dot-pinpoint-dot-chromeperf.appspot.com/

BUG=catapult:#3168

Review-Url: https://codereview.chromium.org/2710293002